display each pokemon with each url
