+++ src/component/pokeInfo/PokeInfo.module.css
@@ -1,0 +1,3 @@
+.PokeInfo {
+  margin: 0;
+}
